Move from prisma1 to prisma2
