+++ .gitignore
@@ -2,3 +2,5 @@
 .next
 node_modules
 .env
+*.log
+backup.sql
