--- backend/context.ts
@@ -1,7 +1,0 @@
-import { Prisma } from './generated/prisma-client'
-import { IncomingMessage } from 'http'
-
-export interface Context {
-  prisma: Prisma
-  request: IncomingMessage
-}
