--- backend/prisma/prisma.yml
@@ -1,16 +1,0 @@
-# Defines your models, each model is mapped to the database as a table.
-datamodel: datamodel.graphql
-
-# Specifies the language and directory for the generated Prisma client.
-generate:
-  - generator: typescript-client
-    output: backend/generated/prisma-client/
-
-# Ensures Prisma client is re-generated after a datamodel change.
-hooks:
-  post-deploy:
-    - prisma generate -p backend/prisma/prisma.yml
-    - nexus-prisma-generate --project backend/prisma/prisma.yml --client backend/generated/prisma-client --output backend/generated/nexus-prisma
-
-endpoint: ${env:PRISMA_ENDPOINT}
-secret: ${env:PRISMA_SECRET}
