--- backend/resolvers/Query.ts
@@ -1,8 +1,0 @@
-import { prismaObjectType } from 'nexus-prisma'
-
-export const Query = prismaObjectType({
-  name: 'Query',
-  definition(t) {
-    t.prismaFields([])
-  },
-})
