--- backend/utils/constants.ts
@@ -1,13 +1,0 @@
-export const constants = {
-  paymentStatus: {
-    PENDING: 'PENDING',
-    PAID: 'PAID',
-    FAILED: 'FAILED',
-    INSUFFICIENT: 'INSUFFICIENT',
-  },
-  successfulStates: [
-    'success',
-    'success_money_on_account',
-    'sent_to_organization',
-  ],
-}
