--- backend/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './constants'
-export * from './transactions'
-export * from './offers'
-export * from './helpers'
