+++ emails/templates/basicTemplate.mjml
