+++ emails/templates/linkCreated.ts
