+++ nexus/configuration.ts
