+++ nexus/types/Transactions.ts
@@ -1,42 +1,44 @@
-import { prismaObjectType, prismaExtendType } from 'nexus-prisma'
-import { stringArg, idArg, intArg } from 'nexus'
+import { objectType, extendType, stringArg, intArg, idArg } from 'nexus'
 import fetch from 'node-fetch'
-import * as querystring from 'querystring'
+import { encode } from 'querystring'
 import { vokativ } from 'vokativ'
-import { capitalize } from '../utils'
 
-import { constants, isTransactionReserved, getSimplyfiedState } from '../utils'
-import { TransactionStatus } from '../generated/prisma-client'
-import { sendEmail } from '../emails'
+import {
+  constants,
+  isTransactionReserved,
+  getSimplyfiedState,
+  capitalize,
+} from '../utils/helpers'
+import { TransactionStatus } from '@prisma/photon'
+import { sendEmail } from '../../emails'
 
 const { PENDING, PAID } = constants.paymentStatus
 
-export const Transaction = prismaObjectType({
+export const Transaction = objectType({
   name: 'Transaction',
-  definition(t) {
-    t.prismaFields({
-      pick: [
-        'id',
-        'createdAt',
-        'firstName',
-        'lastName',
-        'donatedAmount',
-        'amount',
-        'status',
-        'offer',
-      ],
-    })
+  definition: t => {
+    t.model.id()
+    // t.model.createdAt()
+    t.model.firstName()
+    t.model.lastName()
+    t.model.donatedAmount()
+    t.model.amount()
+    t.model.status()
+    t.model.offer()
   },
 })
 
-export const TransactionQuery = prismaExtendType({
+export const TransactionQuery = extendType({
   type: 'Query',
-  definition(t) {
+  definition: t => {
     t.field('recentTransactions', {
       type: 'Transaction',
       list: true,
-      resolve: async (_, _args, { prisma }) => {
-        return prisma.transactions({ where: { status: 'PAID' }, last: 10 })
+      resolve: async (_, {}, { photon }) => {
+        return photon.transactions.findMany({
+          where: { status: 'PAID' },
+          last: 10,
+        })
       },
     })
 
@@ -45,34 +47,12 @@ export const TransactionQuery = prismaExtendType({
       args: {
         id: idArg({ required: true }),
       },
-      resolve: async (_, { id }, { prisma }) => {
-        const currentTransaction = (await prisma.transaction({ id }).$fragment(`
-          fragment transactionApiValues on Transaction {
-            id
-            status
-            email
-            firstName
-            lastName
-            comment
-            amount
-            offer {
-              createdAt
-              firstName
-              lastName
-              email
-              name
-              price
-              beneficator {
-                name
-                organizationId
-                projectId
-                apiId
-                apiSecret
-              }
-            }
-          }
-        `)) as any
-
+      resolve: async (_, { id }, { photon }) => {
+        const currentTransaction = await photon.transactions.findOne({
+          where: { id },
+          include: { offer: { include: { beneficator: true } } },
+        })
+        
         const {
           createdAt,
           price,
@@ -89,7 +69,7 @@ export const TransactionQuery = prismaExtendType({
           },
         } = currentTransaction.offer
 
-        const qs = querystring.encode({
+        const qs = encode({
           fromPledgedDate: JSON.stringify(createdAt).slice(1, 11),
           projectId: projectId,
           apiId,
@@ -127,7 +107,7 @@ export const TransactionQuery = prismaExtendType({
         const isStatusChanged = currentTransaction.status !== newSimplyfiedSate
 
         // update transaction status before emails will fail somehow
-        const updatedTransactionStatus = await prisma.updateTransaction({
+        const updatedTransactionStatus = await photon.transactions.update({
           data: {
             status: newSimplyfiedSate,
             donatedAmount: realDonatedAmount,
@@ -191,7 +171,7 @@ export const TransactionQuery = prismaExtendType({
   },
 })
 
-export const TransactionMutation = prismaExtendType({
+export const TransactionMutation = extendType({
   type: 'Mutation',
   definition(t) {
     t.field('createTransaction', {
@@ -207,18 +187,12 @@ export const TransactionMutation = prismaExtendType({
       resolve: async (
         _,
         { firstName, lastName, email, comment, offerId, amount },
-        { prisma }
+        { photon }
       ) => {
-        const offer = (await prisma.offer({ id: offerId }).$fragment(`
-          fragment OfferTransaction on Offer {
-            amount
-            transactions {
-              amount
-              status
-              createdAt
-            }
-          }
-        `)) as any
+        const offer = await photon.offers.findOne({
+          where: { id: offerId },
+          select: { active: true, amount: true, transactions: true },
+        })
 
         const usedAmount = offer.transactions
           .filter(
@@ -233,13 +207,15 @@ export const TransactionMutation = prismaExtendType({
           throw new Error("Can't create transaction")
         }
 
-        return await prisma.createTransaction({
-          firstName,
-          lastName,
-          email,
-          comment,
-          amount,
-          offer: { connect: { id: offerId } },
+        return await photon.transactions.create({
+          data: {
+            firstName,
+            lastName,
+            email,
+            comment,
+            amount,
+            offer: { connect: { id: offerId } },
+          },
         })
       },
     })
