+++ nexus/types/index.ts
@@ -1,0 +1,8 @@
+export * from './User'
+export * from './Adress'
+export * from './AuthPayload'
+export * from './File'
+export * from './Gallery'
+export * from './Organizations'
+export * from './Offer'
+export * from './Transactions'
