+++ package.json
@@ -15,11 +15,12 @@
   },
   "pre-commit": "lint-staged",
   "resolutions": {
-    "graphql": "14.2.1"
+    "graphql": "14.5.8"
   },
   "dependencies": {
     "@apollo/react-hooks": "^3.0.1",
     "@apollo/react-ssr": "^3.0.1",
+    "@prisma/photon": "^2.0.0-preview019",
     "@types/airtable": "^0.5.6",
     "@types/jsonwebtoken": "^8.3.5",
     "airtable": "^0.7.2",
@@ -35,10 +36,11 @@
     "copy-to-clipboard": "^3.2.0",
     "dotenv": "^8.2.0",
     "formik": "^2.0.3",
-    "graphql": "14.2.1",
+    "graphql": "^14.5.8",
     "graphql-middleware": "^4.0.2",
     "graphql-shield": "^7.0.2",
     "graphql-tag": "^2.10.1",
+    "graphql-tools": "^4.0.6",
     "https-proxy-agent": "^2.2.1",
     "isomorphic-unfetch": "^2.0.0",
     "jsonwebtoken": "^8.5.1",
@@ -46,8 +48,8 @@
     "md5": "^2.2.1",
     "mjml": "^4.5.1",
     "next": "^9.1.4",
-    "nexus": "0.11.6",
-    "nexus-prisma": "0.3.7",
+    "nexus": "^0.12.0-rc.5",
+    "nexus-prisma": "^0.6.1",
     "nexus-prisma-generate-huvik": "^0.3.7",
     "nodemailer": "^6.3.1",
     "prisma-client-lib": "^1.34.10",
@@ -71,6 +73,7 @@
     "@types/cookie": "^0.3.3",
     "@types/graphql": "^14.5.0",
     "@types/lodash.throttle": "^4.1.6",
+    "@types/md5": "^2.1.33",
     "@types/node": "^12.7.3",
     "@types/nodemailer": "^6.2.2",
     "@types/react": "^16.9.2",
